standalone submodules working?
