+++ LowerTriangularArrays/Project.toml
@@ -5,6 +5,7 @@ version = "0.1.0"
 
 [deps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
+Architectures = "c5c4acf7-95e6-4f60-ae13-83e51f04c294"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 GPUArrays = "0c68f7d7-f131-5f86-a1c3-88cf8149b2d7"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
@@ -15,4 +16,6 @@ Adapt = "4.3.0"
 DocStringExtensions = "0.9.5"
 GPUArrays = "11.2.3"
 KernelAbstractions = "0.9.38"
-LinearAlgebra = "1.11.0"
+
+[sources.Architectures]
+path = "../Architectures"
