+++ Utils/Project.toml
@@ -4,6 +4,7 @@ authors = ["SpeedyWeather contributors"]
 version = "0.1.0"
 
 [deps]
+Architectures = "c5c4acf7-95e6-4f60-ae13-83e51f04c294"
 ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 ConstructionBase = "187b0558-2788-49d3-abe0-74a17ed4e7c9"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
@@ -15,8 +16,10 @@ ModelParameters = "4744a3fa-6c31-4707-899e-a3298e4618ad"
 [compat]
 ComponentArrays = "0.15.29"
 ConstructionBase = "1.6.0"
-Dates = "1.11.0"
 DocStringExtensions = "0.9.5"
 DomainSets = "0.7.16"
 MacroTools = "0.5.16"
 ModelParameters = "0.4.5"
+
+[sources.Architectures]
+path = "../Architectures"
